requirements.txt
aldryn-bootstrap3
cmsplugin_iframe2
django-cms-articles
